Prebuilts:Prebuilt images for gummiboot, gnu-efi and uefi-shim.
BZ: 133642

The projects gummiboot, gnu-efi and uefi-shim are not part of
android build system. Each of these projects have a script
(generate-prebuilts.sh) which can be run to generate the respective
executable images. The script is run and the images are included
as part of prebuilts so that the images can be used directly
instead of these projects being built everytime a build command is run.
If any code changes are to be made in these projects, the script shall
be run and updated prebuilts shall be used for the changes to reflect.

Change-Id: Ib1980b336c056cb10bb016566717b0ed04c51e10
Signed-off-by: Ravindranath D <ravindranathx.doddi@intel.com>
Signed-off-by: Mohanraj S <mohanrajx.sakkaraiappan@intel.com>

Added support to extract processorfamily info

Signed-off-by: Sharada <sharadaX.palasamudram.ashok.kumar@intel.com>