Prebuilts:Prebuilt images for gummiboot, gnu-efi and uefi-shim.
The projects gummiboot, gnu-efi and uefi-shim are not part of
android build system. Each of these projects have a script
(generate-prebuilts.sh) which can be run to generate the respective
executable images. The script is run and the images are included
as part of prebuilts so that the images can be used directly
instead of these projects being built everytime a build command is run.
If any code changes are to be made in these projects, the script shall
be run and updated prebuilts shall be used for the changes to reflect.

Change-Id: I1477649bcdd6ba6f09b80895275ecbd78c3d04ae
Signed-off-by: Ravindranath D <ravindranathx.doddi@intel.com>