linux-x86_64: Add pre-built efitools to support baytrail 64-bit KitKat
Issue: BMDA-384

Added x86 64-bit pre-built efitools to support baytrail 64-bit KitKat.
efitools project is not part of android build system. A script
(generate-prebuilts.sh) can be run to generate the executable images.
These images are included as part of prebuilts directly instead of
efitools project being built every time a build command is run. If any
code changes are to be made in efitools projects, the script shall be run
and updated efitools shall be used for the changes to reflect.

Change-Id: I0ff177de7ede96b1fdf4d50839f301167d826093
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>